+++ CHANGELOG.md
@@ -14,6 +14,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - Write message to log if media proxy is enabled.
 - Don't allow `.+-` in names of local custom emojis.
+- Don't allow local custom emojis with `image/apng` media type.
+- Allow remote custom emojis with `image/avif` and `image/jpeg` media types.
 
 ## [4.7.0] - 2025-08-01